GO Fest 2021 3D assets and structure refactor
As more Pokemon are being pushed to the new system, we have updated the repo to have a new 'Pokemon New' folder under the 3D assets folder.

All new 3D assets (costumes, forms, etc) will be added under this new system. As mentioned in a previous commit, we won't be breaking out costume mons anymore as it is too much work to maintain, but all costumes will be within their main folder, but with automatically generated names.